refactor(#zimic): abstract buffer utilities
