+++ packages/zimic/src/utils/buffers.ts
@@ -1,0 +1,29 @@
+export function convertReadableStreamToBuffer(stream: ReadableStream<Uint8Array>): Promise<Buffer> {
+  const reader = stream.getReader();
+  const chunks: Uint8Array[] = [];
+
+  return new Promise((resolve) => {
+    function pump() {
+      void reader.read().then(({ done, value }) => {
+        if (done) {
+          resolve(Buffer.concat(chunks));
+          return;
+        }
+
+        chunks.push(value);
+        pump();
+      });
+    }
+
+    pump();
+  });
+}
+
+export function convertBufferToReadableStream(buffer: Buffer): ReadableStream<Uint8Array> {
+  return new ReadableStream({
+    start(controller) {
+      controller.enqueue(buffer);
+      controller.close();
+    },
+  });
+}
